source/requirements.txt
requests==2.31.0
atlassian_python_api==3.38.0
